start optimizing actual tokenizer

+++ WordPieceTokenizer/src/Token.cs
@@ -3,4 +3,4 @@
 /// <summary>
 /// Token with the id from vocabulary and its start and end positions
 /// </summary>
-public record struct Token(long Id, int Start, int End);
+public record struct Token(int Id, int Start, int End);
